Add name to Hall of Fame
Just stumbled on these docs by chance. Great to see that stuff is being documented properly. And I love to see a hall of fame with some famous names on it. And I wanna take the opportunity to be un-Swedish about the whole thing and enter myself into the list. I hope that's alright. :) I also fixed a typo in Åke's name.